A_Deep_Research_Agent/requirements.txt
google-genai>=0.2.0
google-cloud-aiplatform>=1.38.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
python-dotenv>=1.0.0
